userCheckMapper.xml 삭제 (userMapper.xml로 통합)
userCheckMapper.xml 삭제 (userMapper.xml로 통합)

userCheckMapper 삭제 (userMapper로 통합)

userCheckService 삭제 (userService로 통합)

userCheckServiceImpl 삭제 (userServiceImpl로 통합)

UserCheckServiceTests 삭제 (UserServiceTests)로 통합
