--- src/main/java/com/ncncn/mapper/UserCheckMapper.java
@@ -1,22 +1,0 @@
-package com.ncncn.mapper;
-
-import com.ncncn.domain.*;
-import com.ncncn.domain.pagination.UserCheckCriteria;
-
-import java.util.List;
-
-public interface UserCheckMapper {
-
-    public List<UserVO> getUserList();
-
-    public List<UserVO> getListWithPaging(UserCheckCriteria cri);
-
-    public int getTotalCount(UserCheckCriteria cri);
-
-    public UserDetailCheckVO getUserDetail(int userId);
-
-    public void updateMemo(UserMemoVO memo);
-
-    public void updateStatus(UserStatusVO status);
-
-}
