+++ src/main/java/com/ncncn/service/UserService.java
@@ -1,11 +1,28 @@
 package com.ncncn.service;
 
+import com.ncncn.domain.UserDetailCheckVO;
+import com.ncncn.domain.UserMemoVO;
+import com.ncncn.domain.UserStatusVO;
 import com.ncncn.domain.UserVO;
+import com.ncncn.domain.pagination.UserCheckCriteria;
+
+import java.util.List;
 
 public interface UserService {
 
-    public UserVO readbyId(int userId);
+    UserVO readbyId(int userId);
 
 	int countRecentlyRegister();
+
+
+    List<UserVO> getUserList(UserCheckCriteria cri);
+
+    int getTotal(UserCheckCriteria cri);
+
+    UserDetailCheckVO getUserDetail(int userId);
+
+    void updateMemo(UserMemoVO memo);
+
+    void updateStatus(UserStatusVO status);
 
 }
